+++ CHANGELOG.md
@@ -13,6 +13,9 @@ We used to have incremental versioning before `0.1.0`.
 - Updates `flake8` to `3.7+`
 - Forbids to use redundant `+`, `~`, `not`, and `-` operators before numbers
 - Forbids to use complex default values
+- Forbids to use anything rather than names in `for` loop vars definitions
+- Forbids to use anything rather than names in `with` block vars definitions
+- Forbids to use anything rather than names in comprehension vars definitions
 
 ### Bugfixes